Add more MessageEvent functionality and tests.
This adds origin, lastEventId, source, and ports to MessageEvent, and
adds a set of related unit tests for MessageEvent and
ExtendableMessageEvent.

To support the unit tests, the TestWithJavaScript helper classes have
been changed to use a virtual accessor for the web::Context instead of
the script::GlobalEnvironment, as that still allows access to the latter
and also allows access to the WindowOrWorkerGlobalScope needed for
testing the 'source' property. To still correctly initialize the window_
for the web::testing tests with the Window global scope a call to
InitializeWindow() had to be added.

Also, the EvaluateScript() overload that returns a
script::ValueHandleHolder::Reference was added to the TestWithJavaScript
helpers to allow the unit tests to test with the data property set to a
JS value.

Also renamed GetWorkerTypes() to GetWebTypes() for the web::testing test
case instantiation.

The above changes somehow also triggered some previously unseen compiler
warnings in script/v8c, and some other places that are also fixed in
this CL.

b/236294349

NOKEYCHECK=True
GitOrigin-RevId: 90af64cbba9cc6860d2ac8cd828fa4a6cf9d05f0

+++ cobalt/script/v8c/helpers.h
@@ -15,6 +15,8 @@
 #ifndef COBALT_SCRIPT_V8C_HELPERS_H_
 #define COBALT_SCRIPT_V8C_HELPERS_H_
 
+#include <string>
+
 #include "starboard/common/string.h"
 #include "v8/include/v8.h"
 
@@ -35,7 +37,7 @@ inline v8::Local<v8::String> NewInternalString(v8::Isolate* isolate,
   }
   return v8::String::NewFromOneByte(
              isolate, reinterpret_cast<const uint8_t*>(string),
-             v8::NewStringType::kInternalized, strlen(string))
+             v8::NewStringType::kInternalized, static_cast<int>(strlen(string)))
       .ToLocalChecked();
 }
